+++ src/unittests/Xml_Test.cxx
@@ -21,7 +21,7 @@ TEST(XmlTests, BasicTests)
     
     EXPECT_EQ(8,toks.Count());
     
-    for(size_t i=0;i<toks.Count();i++)
+    for(int i=0;i<toks.Count();i++)
     {
         System::Console::WriteLine(toks[i].ToString());
     }
